chore: update VIP number
