+++ simulations/vip-157/vip-157-testnet/abi/ResilientOracle_ABI.json
